Re-organize module tests
